expose interfaces to __init__.py files
