+++ botstory/integrations/fb/__init__.py
@@ -1,0 +1,1 @@
+from .messenger import FBInterface
