+++ botstory/integrations/mockdb/__init__.py
@@ -1,16 +1,1 @@
-class MockDB:
-    def __init__(self):
-        self.session = None
-        self.user = None
-
-    async def get_session(self, **kwargs):
-        return self.session
-
-    async def set_session(self, session):
-        self.session = session
-
-    async def get_user(self, **kwargs):
-        return self.user
-
-    async def set_user(self, user):
-        self.user = user
+from .db import MockDB
