+++ botstory/middlewares/any/__init__.py
@@ -1,0 +1,1 @@
+from .any import Any, AnyOf
